build: Support Maven and Eclipse
Add pom.xml to compile PrologCafe using `mvn package`. Currently this
is a pretty big hack.  The base builtins (that are human written)
had to be moved to the builtin package, and a symlink is required
to setup the Java package structure correctly for the Eclipse IDE.

Most of the build still runs the standard Makefile, but is invoked
from Maven's antrun plugin. This way we don't need to immediately
rewrite the parts of the build that convert Prolog to Java.

At the end the Maven artifact PrologCafe should be logically
identical to the plcafe.jar that the Makefile system builds.

A better Maven build would split the compiler into its own JAR,
as this is easily half of the build output.  Unfortunately any
downstream consumer probably needs the compiler anyway, so its
probably not worth trying to split out.