Tighten up the Trail related code for VariableTerm
Mark some of this code final, to ensure the compiler and JIT
know exactly what they can inline at runtime, since there is
no replacement possible.

Shorten the chain from the Trail to the ChoicePointStack. This
way the VariableTerm can find the current frame's timestamp in
less redirections and avoid some cost associated with binding
a variable to a value.