Remove incorrect copyright header from Google
My Eclipse was misconfigured years ago and wrote this header
on ChoicePointFrame when the class was refactored out of other
code that came from the original Prolog Cafe. Google never had
the intention of reserving all rights; the file is under the dual
GPL+EPL license used by the project.

Change-Id: I165c6710bfd6cb47c8816799ead574de0d8c8139