Use Java based table-switch for type switches
Predicates that use different clauses for different types of terms use
a type switch on the first argument/register.  Rewrite this as a
proper table switch in Java so the JVM can emit an efficient switch
statement.  Use a virtual method and lightweight ints enumerated from
0 to N in the order the switch was previously written in to get an
efficient implementation for the JIT.

Change-Id: If7274ec8776479afe70bbec0882768d7a28b4c64